+++ crowdgezwitscher/facebook/urls_api.py
@@ -2,7 +2,7 @@
 from django.conf.urls import url
 from rest_framework.urlpatterns import format_suffix_patterns
 
-from . import views
+from . import views_api as views
 
 app_name = 'facebook_api'
 urlpatterns_raw = (
