+++ crowdgezwitscher/facebook/views.py
@@ -1,15 +1,11 @@
-from rest_framework import generics
-
 from django.contrib.auth.mixins import PermissionRequiredMixin
 from django.views.generic.list import ListView
 from django.views.generic.detail import DetailView
 from django.views.generic.edit import CreateView, UpdateView, DeleteView
 from django.urls import reverse_lazy
 
-from base.models import MapObjectFilterBackend
 from facebook.forms import FacebookPageForm
 from facebook.models import FacebookPage
-from facebook.serializers import FacebookPageSerializer, FacebookPageSerializerShortened
 
 
 class FacebookPageListView(PermissionRequiredMixin, ListView):
@@ -46,15 +42,3 @@ class FacebookPageDelete(PermissionRequiredMixin, DeleteView):
     template_name = 'facebook/delete.html'
     success_url = reverse_lazy('facebook:list')
     context_object_name = 'page'
-
-
-# API Views
-class FacebookPageAPIList(generics.ListAPIView):
-    queryset = FacebookPage.objects.filter(active=True)
-    serializer_class = FacebookPageSerializerShortened
-    filter_backends = (MapObjectFilterBackend,)
-
-
-class FacebookPageAPIDetail(generics.RetrieveAPIView):
-    queryset = FacebookPage.objects.filter(active=True)
-    serializer_class = FacebookPageSerializer
